fix: token generation moved to backend
squash

clean up some console
