+++ apps/web/app/ts_samples/echo_fast/page.tsx
@@ -1,4 +1,4 @@
-import { generate_random_token } from "../../api_handler";
+import { generate_random_token } from "../../actions/token";
 import EchoFastContent from "./content";
 
 export default async function EchoFast() {
