+++ apps/web/app/ts_samples/multi_track_receivers/page.tsx
@@ -1,4 +1,4 @@
-import { generate_random_token } from "../../api_handler";
+import { generate_random_token } from "../../actions/token";
 import Content from "./content";
 
 export default async function MultiTrackReceivers() {
