+++ turbo.json
@@ -4,7 +4,7 @@
   "pipeline": {
     "build": {
       "dependsOn": ["^build"],
-      "env": ["NEXT_PUBLIC_GATEWAYS", "NEXT_PUBLIC_APP_TOKEN"],
+      "env": ["NEXT_PUBLIC_GATEWAYS", "APP_SECRET"],
       "outputs": [".next/**", "!.next/cache/**"]
     },
     "lint": {
